+++ src/commands/hardis/org/purge/flow.ts
@@ -2,8 +2,9 @@
 import { flags, SfdxCommand } from '@salesforce/command';
 import { Messages, SfdxError } from '@salesforce/core';
 import { AnyJson } from '@salesforce/ts-types';
+import * as c from 'chalk';
 import * as columnify from 'columnify';
-import * as sfdx from 'sfdx-node';
+import { execSfdxJson, uxLog } from '../../../../common/utils';
 
 // Initialize Messages with the current plugin directory
 Messages.importMessagesDirectory(__dirname);
@@ -13,13 +14,12 @@ Messages.importMessagesDirectory(__dirname);
 const messages = Messages.loadMessages('sfdx-hardis', 'org');
 
 export default class OrgPurgeFlow extends SfdxCommand {
-
   public static title = 'Purge Flow versions';
 
   public static description = messages.getMessage('orgPurgeFlow');
 
   public static examples = [
-  `$ sfdx hardis:org:purge:flow --targetusername nicolas.vuillamy@gmail.com
+    `$ sfdx hardis:org:purge:flow --targetusername nicolas.vuillamy@gmail.com
   Found 1 records:
   ID                 MASTERLABEL VERSIONNUMBER DESCRIPTION  STATUS
   30109000000kX7uAAE TestFlow    2             test flowwww Obsolete
@@ -29,7 +29,7 @@ export default class OrgPurgeFlow extends SfdxCommand {
   ID                 MASTERLABEL VERSIONNUMBER DESCRIPTION  STATUS
   30109000000kX7uAAE TestFlow    2             test flowwww Obsolete
   `,
-  `$ sfdx hardis:org:purge:flow --targetusername nicolas.vuillamy@gmail.com --status "Obsolete,Draft,InvalidDraft --name TestFlow"
+    `$ sfdx hardis:org:purge:flow --targetusername nicolas.vuillamy@gmail.com --status "Obsolete,Draft,InvalidDraft --name TestFlow"
   Found 4 records:
   ID                 MASTERLABEL VERSIONNUMBER DESCRIPTION  STATUS
   30109000000kX7uAAE TestFlow    2             test flowwww Obsolete
@@ -45,13 +45,41 @@ export default class OrgPurgeFlow extends SfdxCommand {
 
   protected static flagsConfig = {
     // flag with a value (-n, --name=VALUE)
-    prompt: flags.boolean({char: 'z', default: true, allowNo: true,  description: messages.getMessage('prompt')}),
-    name: flags.string({char: 'n', description: messages.getMessage('nameFilter')}),
-    status: flags.string({char: 's', default: 'Obsolete', description: messages.getMessage('statusFilter')}),
-    allowpurgefailure: flags.boolean({char: 'f', default: true, allowNo: true,  description: messages.getMessage('allowPurgeFailure')}),
-    sandbox: flags.boolean({ default: false, description: messages.getMessage('sandboxLogin')}),
-    instanceurl: flags.string({char: 'r', default: 'https://login.saleforce.com',  description: messages.getMessage('instanceUrl')}),
-    debug: flags.boolean({char: 'd', default: false, description: messages.getMessage('debugMode')})
+    prompt: flags.boolean({
+      char: 'z',
+      default: true,
+      allowNo: true,
+      description: messages.getMessage('prompt')
+    }),
+    name: flags.string({
+      char: 'n',
+      description: messages.getMessage('nameFilter')
+    }),
+    status: flags.string({
+      char: 's',
+      default: 'Obsolete',
+      description: messages.getMessage('statusFilter')
+    }),
+    allowpurgefailure: flags.boolean({
+      char: 'f',
+      default: true,
+      allowNo: true,
+      description: messages.getMessage('allowPurgeFailure')
+    }),
+    sandbox: flags.boolean({
+      default: false,
+      description: messages.getMessage('sandboxLogin')
+    }),
+    instanceurl: flags.string({
+      char: 'r',
+      default: 'https://login.saleforce.com',
+      description: messages.getMessage('instanceUrl')
+    }),
+    debug: flags.boolean({
+      char: 'd',
+      default: false,
+      description: messages.getMessage('debugMode')
+    })
   };
 
   // Comment this out if your command does not require an org username
@@ -66,11 +94,14 @@ export default class OrgPurgeFlow extends SfdxCommand {
   /* jscpd:ignore-end */
 
   public async run(): Promise<AnyJson> {
-    const prompt = (this.flags.prompt === false) ? false : true;
-    const statusFilter = (this.flags.status) ? this.flags.status.split(',') : ['Obsolete'];
+    const prompt = this.flags.prompt === false ? false : true;
+    const statusFilter = this.flags.status
+      ? this.flags.status.split(',')
+      : ['Obsolete'];
     const nameFilter = this.flags.name || null;
-    const allowPurgeFailure = (this.flags.allowpurgefailure === false) ? false : true;
-    const debug = this.flags.debug || false ;
+    const allowPurgeFailure =
+      this.flags.allowpurgefailure === false ? false : true;
+    const debugMode = this.flags.debug || false;
 
     // Check we don't delete active Flows
     if (statusFilter.includes('Active')) {
@@ -78,9 +109,10 @@ export default class OrgPurgeFlow extends SfdxCommand {
     }
 
     // Build query with name filter if sent
-    let query = `SELECT Id,MasterLabel,VersionNumber,Status,Description FROM Flow WHERE Status IN ('${statusFilter.join("','")}')`;
+    let query = `SELECT Id,MasterLabel,VersionNumber,Status,Description FROM Flow WHERE Status IN ('${statusFilter.join(
+      "','"
+    )}')`;
     if (statusFilter) {
-
     }
     if (nameFilter) {
       query += ` AND MasterLabel LIKE '${nameFilter}%'`;
@@ -90,62 +122,100 @@ export default class OrgPurgeFlow extends SfdxCommand {
     const username = this.org.getUsername();
 
     // const flowQueryResult = await conn.query<Flow>(query,{ });
-    const flowQueryResult = await sfdx.data.soqlQuery({
-      query,
-      targetusername: username,
-      usetoolingapi: true,
-      _quiet: !debug,
-      _rejectOnError: true
+    const flowQueryCommand = 'sfdx force:data:soql:query ' + ` -q "${query}"` +
+    ` --targetusername ${username}` + ' --usetoolingapi';
+    const flowQueryRes = await execSfdxJson(flowQueryCommand, this, {
+      output: false,
+      debug: debugMode,
+      fail: true
     });
 
     // Check empty result
-    if (!flowQueryResult.records || flowQueryResult.records.length <= 0) {
+    if (!flowQueryRes.records || flowQueryRes.records.length <= 0) {
       const outputString = `[sfdx-hardis] No matching Flow records found with query ${query}`;
       this.ux.log(outputString);
-      return { deleted: [] , outputString};
+      return { deleted: [], outputString };
     }
 
     // Simplify results format & display them
-    const records = flowQueryResult.records.map(record => {
-      return { Id: record.Id, MasterLabel: record.MasterLabel, VersionNumber: record.VersionNumber,
-         Description: record.Description, Status: record.Status };
+    const records = flowQueryRes.records.map(record => {
+      return {
+        Id: record.Id,
+        MasterLabel: record.MasterLabel,
+        VersionNumber: record.VersionNumber,
+        Description: record.Description,
+        Status: record.Status
+      };
     });
-    this.ux.log(`[sfdx-hardis] Found ${records.length} records:\n${columnify(records)}`);
+    this.ux.log(
+      `[sfdx-hardis] Found ${c.bold(records.length)} records:\n${c.grey(
+        columnify(records)
+      )}`
+    );
 
     // Perform deletion
     const deleted = [];
     const deleteErrors = [];
     if (
       !prompt ||
-      await this.ux.confirm(`[sfdx-hardis] Are you sure you want to delete this list of records in ${this.org.getUsername()} (y/n)?`)
-      ) {
-        for (const record of records) {
-          const deleteResult = await sfdx.data.recordDelete({
-            sobjecttype: 'Flow',
-            sobjectid: record.Id,
-            targetusername: username,
-            usetoolingapi: true,
-            _quiet: debug
-          });
-          if ((deleteResult && deleteResult.success !== true) || deleteResult == null) {
-            this.ux.error(`[sfdx-hardis] Unable to perform deletion request: ${JSON.stringify(deleteResult || {})}`);
-            deleteErrors.push(deleteResult);
-          }
-          deleted.push(record);
+      (await this.ux.confirm(
+        c.bold(
+          `[sfdx-hardis] Are you sure you want to delete this list of records in ${c.green(
+            this.org.getUsername()
+          )} (y/n)?`
+        )
+      ))
+    ) {
+      for (const record of records) {
+        const deleteCommand =
+          'sfdx force:data:record:delete' +
+          ' --sobjecttype Flow' +
+          ` --sobjectid ${record.Id}` +
+          ` --targetusername ${username}` +
+          ' --usetoolingapi';
+        const deleteRes = await execSfdxJson(deleteCommand, this, {
+          fail: false,
+          output: false,
+          debug: debugMode
+        });
+        if (!(deleteRes.status === 0)) {
+          this.ux.error(
+            c.red(
+              `[sfdx-hardis] Unable to perform deletion request: ${JSON.stringify(
+                deleteRes
+              )}`
+            )
+          );
+          deleteErrors.push(deleteRes);
         }
-    }
-
-    if (deleteErrors.length > 0) {
-      const errMsg = `[sfdx-hardis] There are been errors while deleting ${deleteErrors.length} records: \n${JSON.stringify(deleteErrors)}`;
-      if (allowPurgeFailure) {
-        console.warn(errMsg);
-      } else {
-        throw new SfdxError(`There are been errors while deleting ${deleteErrors.length} records: \n${JSON.stringify(deleteErrors)}`);
+        deleted.push(record);
       }
     }
 
-    const summary = (deleted.length > 0) ? `[sfdx-hardis] Deleted the following list of records:\n${columnify(deleted)}` : '[sfdx-hardis] No record deleted';
-    this.ux.log(summary);
+    if (deleteErrors.length > 0) {
+      const errMsg = `[sfdx-hardis] There are been errors while deleting ${
+        deleteErrors.length
+      } records: \n${JSON.stringify(deleteErrors)}`;
+      if (allowPurgeFailure) {
+        uxLog(this, c.yellow(errMsg));
+      } else {
+        throw new SfdxError(
+          c.yellow(
+            `There are been errors while deleting ${
+              deleteErrors.length
+            } records: \n${JSON.stringify(deleteErrors)}`
+          )
+        );
+      }
+    }
+
+    const summary =
+      deleted.length > 0
+        ? `[sfdx-hardis] Deleted the following list of records:\n${columnify(
+            deleted
+          )}`
+        : '[sfdx-hardis] No record to delete';
+    this.ux.log(c.green(summary));
     // Return an object to be displayed with --json
     return { orgId: this.org.getOrgId(), outputString: summary };
   }